[DomTree] Replace ChildrenGetter with GraphTraits over GraphDiff.
This replaces the ChildrenGetter inside the DominatorTree with
GraphTraits over a GraphDiff object, an object which encapsulated the
view of the previous CFG.
This also simplifies the extentions in clang which use DominatorTree, as
GraphDiff also filters nullptrs.

Re-land 2f28169767395a130442befae6e355b025dda53c after moving CFGDiff.h
to Support.

Differential Revision: https://reviews.llvm.org/D77341