Re-land [Debug][CodeView] Emit fully qualified names for globals
This reverts commit 13284c4110df7a583fa4d4e3368f7ee2c5677d7d.

Fixed an issue with pointers to members based on typedefs. In this case,
LLVM would emit a second UDT. I fixed it by not passing the class type
to getTypeIndex when the base type is not a function type. lowerType
only uses the class type for direct function types. This suggests if we
have a PMF with a function typedef, there may be an issue, but that can
be solved separately.