[IR] Convert profile metadata in createCallMatchingInvoke()
When an invoke instruction is converted to a call its
profile metadata is dropped because it has incompatible
format (see commit 6cc9c7a0cc3f).
This patch adds an attempt to convert profile data to
format of the call instruction. This used to work well
before the commit b3a35b3de06f.

Reviewers: reames
Tags: #llvm
Differential Revision: https://reviews.llvm.org/D82071