Remove Passes dependency on CodeGen
The dependency was introduced in
40d2dad7a519a6283e5f5321b7909f31af562789. The only functional change
from this removal would be the new PM interface for the two codegen
passes. This is not necessary since we don't have codegen pipeline using
new PM yet. This removal is to break the potential circular dependency between
Passes and CodeGen once the codegen begins to gain new PM support.